requirements.txt
requests>=2.31.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.97.0
edge-tts>=6.1.7
moviepy>=1.0.3
pillow>=10.0.0
openai>=1.3.5
google-generativeai>=0.4.0
replicate>=0.21.1
pexels-api>=1.0.1
pixabay-python>=0.5
youtube-upload>=0.3.3
tavily-python>=0.3.3
python-telegram-bot>=20.7
numpy>=1.24.3
ffmpeg-python>=0.2.0
pydub>=0.25.1
feedparser>=6.0.10
wikipedia-api>=0.5.8
huggingface-hub>=0.20.1
python-dotenv>=1.0.0
pytz>=2023.3
schedule>=1.2.0
urllib3>=2.0.4
certifi>=2023.7.22
charset-normalizer>=3.2.0
idna>=3.4
